create a single router file
