+++ wolox-books/src/app/app-routing.module.ts
@@ -1,13 +1,19 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { SignUpComponent } from './screens/Sign-up/sign-up.component';
+import { LoginComponent } from './screens/login/login.component';
 
-const routes: Routes = [
-    {
-        path: '',
-        pathMatch: 'prefix',
-        redirectTo: '/login',
-    }
-]
+const routes: Routes = [{
+    path: '',
+    pathMatch: 'full',
+    redirectTo: '/login',
+}, {
+    path: 'sign-up',
+    component: SignUpComponent
+}, {
+    path: 'login',
+    component: LoginComponent
+}]
 
 @NgModule({
     imports: [RouterModule.forRoot(routes)],
