--- wolox-books/src/app/screens/Sign-up/sign-up-routing.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { SignUpComponent } from './sign-up.component';
-
-const routes: Routes = [{
-  path: 'sign-up',
-  component: SignUpComponent
-}]
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class SignUpRoutingModule { }
