--- wolox-books/src/app/screens/login/login-routing.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { LoginComponent } from './login.component';
-
-const routes: Routes = [{
-  path: 'login',
-  component: LoginComponent
-}];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class LoginRoutingModule { }
